chore(discord): Move import from discord functionality to separatefile
